refactor: functional
