+++ .gitignore
@@ -1,2 +1,1 @@
-*.wav
-*.mp3
+*.code*
