+++ audioGuestBook.service
@@ -1,0 +1,11 @@
+[Unit]
+Description=Rotary Phone Guest Book Project
+After=multi-user.target
+[Service]
+WorkingDirectory=/home/<username>/Desktop/rotaryGuestBook
+User=<username>
+Type=simple
+Restart=always
+ExecStart=/usr/bin/env python3 /home/<username>/Desktop/rotaryGuestBook/rotaryGuestBook.py
+[Install]
+WantedBy=multi-user.target
